@@ -6,7 +6,7 @@ import {
   effect,
   OnDestroy,
 } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 import { Router } from '@angular/router';
 import { CartService } from '../../services/cart/cart.service';
 import { IconComponent } from '../shared/icon/icon.component';
@@ -14,7 +14,7 @@ import { IconComponent } from '../shared/icon/icon.component';
 @Component({
   selector: 'app-cart-added-modal',
   standalone: true,
-  imports: [CommonModule, IconComponent],
+  imports: [IconComponent],
   templateUrl: './cart-added-modal.component.html',
   styleUrl: './cart-added-modal.component.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
