@@ -1,10 +1,10 @@
-import { CommonModule } from '@angular/common';
+
 import { Component, input } from '@angular/core';
 
 @Component({
   selector: 'app-custom-design-preview',
   standalone: true,
-  imports: [CommonModule],
+  imports: [],
   templateUrl: './custom-design-preview.component.html',
   styleUrl: './custom-design-preview.component.scss',
 })
