@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 import { IconComponent } from '../shared/icon/icon.component';
 
 @Component({
   selector: 'app-footer',
   standalone: true,
-  imports: [CommonModule, IconComponent],
+  imports: [IconComponent],
   templateUrl: './footer.component.html',
   styleUrls: ['./footer.component.scss'],
 })
