@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-import { CommonModule } from '@angular/common';
+
 import { ContactFormComponent } from '../../components/contact-form/contact-form.component';
 
 @Component({
   selector: 'app-licensing',
   standalone: true,
-  imports: [CommonModule, ContactFormComponent],
+  imports: [ContactFormComponent],
   templateUrl: './licensing.component.html',
   styleUrls: ['./licensing.component.scss'],
 })
